write policies for coordinates

+++ app/controllers/api/v1/coordinates_controller.rb
@@ -1,14 +1,20 @@
 class Api::V1::CoordinatesController < ApplicationController
+  include ResourceFinder
   before_action :set_coordinate, only: %i[show update destroy]
   skip_before_action :authenticate_request, only: %i[show]
-  include ResourceFinder
+  before_action :authorize_policy
 
   def show
+    authorize @coordinate
+
     render json: @coordinate
   end
 
   def create
     @coordinate = parentable.coordinates.new(coordinate_params)
+
+    authorize @coordinate
+
     if @coordinate.save
       render json: @coordinate, status: :created
     else
@@ -17,6 +23,8 @@ class Api::V1::CoordinatesController < ApplicationController
   end
 
   def update
+    authorize @coordinate
+
     if @coordinate.update(coordinate_params)
       render json: { status: 'Update', data: @coordinate }, status: :ok
     else
@@ -25,6 +33,8 @@ class Api::V1::CoordinatesController < ApplicationController
   end
 
   def destroy
+    authorize @coordinate
+    
     if @coordinate.destroy!
       render json: { status: 'Delete' }, status: :no_content
     else
@@ -40,5 +50,9 @@ class Api::V1::CoordinatesController < ApplicationController
 
   def coordinate_params
     params.permit(:latitude, :longitude)
+  end
+
+  def authorize_policy
+    authorize Coordinate
   end
 end
